+++ CHANGELOG.rst
@@ -5,6 +5,19 @@ Sitehost.Cloud Release Notes
 .. contents:: Topics
 
 
+v1.1.0
+======
+
+Bugfixes
+--------
+
+- Fixed an issue with server module when state is `started` or `stopped`.
+
+New Modules
+-----------
+
+- sitehost.cloud.dns - Manage DNS zones and records
+
 v1.0.0
 ======